Added support for colouring the Xonlinedash.xbe. Updated the Xboxdash 5960 Colourizer.zip an python script to use a UI making it simpler to use.
Updated the Build UIX Lite Skins tool with the new images for xonlinedash and also added the blood red skin so people can tinker with it.

Also added Milenkos script for setting colours via a UI, fixed up an improved.
It now works in a similar way to the above scrip allowing a base colour to be set. Currently only xboxdash.xbe